requirements.txt
# Advanced Driver Safety System - Python Dependencies

# Core computer vision and ML
opencv-python>=4.8.0
dlib>=19.24.0
imutils>=0.5.4
scipy>=1.11.0
numpy>=1.24.0
pillow>=10.0.0

# Hardware interfaces
gpiozero>=2.0
spidev>=3.6
pyserial>=3.5

# Audio and alerts
pygame>=2.5.0
pyaudio>=0.2.13

# Speech recognition
SpeechRecognition>=3.10.0

# Sensor libraries
adxl345-python>=1.0.0

# Email functionality (built-in, no install needed)
# - smtplib
# - email.mime

# System requirements (install via apt):
# sudo apt-get install -y portaudio19-dev python3-pyaudio flac python3-serial

# Installation command:
# pip install --break-system-packages -r requirements.txt
